+++ cookbooks/aws-parallelcluster-computefleet/recipes/install/custom_parallelcluster_node.rb
@@ -50,7 +50,7 @@ unless node['cluster']['install_python_from_internet']
   end
 end
 
-bash "install custom aws-parallelcluster-node" do
+bash "install aws-parallelcluster-node" do
   cwd Chef::Config[:file_cache_path]
   code <<-NODE
     set -e