+++ cookbooks/aws-parallelcluster-shared/attributes/versions.rb
@@ -6,6 +6,13 @@ default['cluster']['python-major-minor-version'] = '3.14'
 # Set to true when testing new Python versions before artifacts are available in S3
 default['cluster']['install_python_from_internet'] = false
 
+# Opt-out: install the aws-parallelcluster-node package from PyPI instead of the
+# official S3 bucket.
+default['cluster']['install_node_from_internet'] = false
+
+# Official ParallelCluster Node Package
+default['cluster']['custom_node_package'] = "#{node['cluster']['s3_url']}/parallelcluster/#{node['cluster']['parallelcluster-node-version']}/node/aws-parallelcluster-node-#{node['cluster']['parallelcluster-node-version']}.tgz"
+
 # ParallelCluster versions
 default['cluster']['parallelcluster-version'] = '3.16.0'
 default['cluster']['parallelcluster-cookbook-version'] = '3.16.0'